Use parseFloat instead of parseInt

+++ bundle.js
@@ -18,10 +18,10 @@ Hooks.once("dragRuler.ready", (SpeedProvider) => {
 
 			if(tokenElevation > 0) {
 				for(var i=0, len= tokenSpeed.length; i<len; i++){
-					if (tokenSpeed[i].id === 'fly') {var baseSpeed = parseInt(tokenSpeed[i].value)}
+					if (tokenSpeed[i].id === 'fly') {var baseSpeed = parseFloat(tokenSpeed[i].value)}
 			 	}
-				if (baseSpeed === undefined) {var baseSpeed = parseInt(tokenSpeed[0].value)}
-			}else if (tokenElevation <= 0){var baseSpeed = parseInt(tokenSpeed[0].value)}
+				if (baseSpeed === undefined) {var baseSpeed = parseFloat(tokenSpeed[0].value)}
+			}else if (tokenElevation <= 0){var baseSpeed = parseFloat(tokenSpeed[0].value)}
 
 
 			for (var i=0, len=conditions.length; i<len; i++) {
@@ -59,25 +59,25 @@ var land = 0; var fly = 0; var swim = 0; var climb = 0; var burrow = 0; var misc
 	if (token.actor.data.type === "character"){
 		const cleanedSpeeds = speedsSorter(token)
 		var land = cleanedSpeeds.land; var fly = cleanedSpeeds.fly; var swim = cleanedSpeeds.swim; var climb = cleanedSpeeds.swim; var burrow = cleanedSpeeds.burrow; var miscSpeed =cleanedSpeeds.misc;
-		var land = parseInt(token.actor.data.data.attributes.speed.total);
+		var land = parseFloat(token.actor.data.data.attributes.speed.total);
 
 	} else if (token.actor.data.type === "npc"){
 		const cleanedSpeeds = speedsSorter(token)
 		var land = cleanedSpeeds.land; var fly = cleanedSpeeds.fly; var swim = cleanedSpeeds.swim; var climb = cleanedSpeeds.swim; var burrow = cleanedSpeeds.burrow; var miscSpeed =cleanedSpeeds.misc;
 		if(token.actor.data.data.attributes.speed.type === "land" || token.actor.data.data.attributes.speed.type === "Land" ){
-			var land = parseInt(token.actor.data.data.attributes.speed.total)
+			var land = parseFloat(token.actor.data.data.attributes.speed.total)
 		} else if (token.actor.data.data.attributes.speed.type === "fly" || token.actor.data.data.attributes.speed.type === "Fly") {
-			var fly = parseInt(token.actor.data.data.attributes.speed.total);
+			var fly = parseFloat(token.actor.data.data.attributes.speed.total);
 		} else if (token.actor.data.data.attributes.speed.type === "swim" || token.actor.data.data.attributes.speed.type === "Swim") {
-			var swim = parseInt(token.actor.data.data.attributes.speed.total);
+			var swim = parseFloat(token.actor.data.data.attributes.speed.total);
 		}	else if (token.actor.data.data.attributes.speed.type === "climb" || token.actor.data.data.attributes.speed.type === "Climb" ) {
-			var climb = parseInt(token.actor.data.data.attributes.speed.total);
+			var climb = parseFloat(token.actor.data.data.attributes.speed.total);
 		} else if (token.actor.data.data.attributes.speed.type === "burrow" || token.actor.data.data.attributes.speed.type === "Burrow") {
-			var burrow = parseInt(token.actor.data.data.attributes.speed.total);
-		} else{var miscSpeed = parseInt(token.actor.data.data.attributes.speed.total)};
+			var burrow = parseFloat(token.actor.data.data.attributes.speed.total);
+		} else{var miscSpeed = parseFloat(token.actor.data.data.attributes.speed.total)};
 
 		if (land === 0 && token.actor.data.data.attributes.speed.value !== undefined) {
-			var miscSpeed = parseInt(token.actor.data.data.attributes.speed.value.match(/\d+/));
+			var miscSpeed = parseFloat(token.actor.data.data.attributes.speed.value.match(/\d+/));
 		};
 
 	} else if (token.actor.data.type === "familiar"){
@@ -88,7 +88,7 @@ var land = 0; var fly = 0; var swim = 0; var climb = 0; var burrow = 0; var misc
 
 
 	} else if (token.actor.data.type === "vehicle"){
-		var land = parseInt(token.actor.data.data.details.speed)
+		var land = parseFloat(token.actor.data.data.details.speed)
 	}
 
 	const speeds = [];
@@ -108,16 +108,16 @@ var land2 = 0; var fly2 = 0; var swim2 = 0; var climb2 = 0; var burrow2 = 0; var
 if (token.actor.data.data.attributes.speed.otherSpeeds !== undefined){
 	for (var i=0, len=token.actor.data.data.attributes.speed.otherSpeeds.length; i<len; i++){
 		if(token.actor.data.data.attributes.speed.otherSpeeds[i].type === "land" || token.actor.data.data.attributes.speed.otherSpeeds[i].type === "Land"){
-			var land2 = parseInt(token.actor.data.data.attributes.speed.otherSpeeds[i].total)
+			var land2 = parseFloat(token.actor.data.data.attributes.speed.otherSpeeds[i].total)
 		} else if (token.actor.data.data.attributes.speed.otherSpeeds[i].type === "fly" || token.actor.data.data.attributes.speed.otherSpeeds[i].type === "Fly") {
-			var fly2 = parseInt(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
+			var fly2 = parseFloat(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
 		} else if (token.actor.data.data.attributes.speed.otherSpeeds[i].type === "swim" || token.actor.data.data.attributes.speed.otherSpeeds[i].type === "Swim") {
-			var swim2 = parseInt(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
+			var swim2 = parseFloat(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
 		}	else if (token.actor.data.data.attributes.speed.otherSpeeds[i].type === "climb" || token.actor.data.data.attributes.speed.otherSpeeds[i].type === "Climb" ) {
-			var climb2 = parseInt(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
+			var climb2 = parseFloat(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
 		} else if (token.actor.data.data.attributes.speed.otherSpeeds[i].type === "burrow" || token.actor.data.data.attributes.speed.otherSpeeds[i].type === "Burrow" ) {
-			var burrow2 = parseInt(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
-		} else{var miscSpeed2 = parseInt(token.actor.data.data.attributes.speed.otherSpeeds[i].total)}
+			var burrow2 = parseFloat(token.actor.data.data.attributes.speed.otherSpeeds[i].total);
+		} else{var miscSpeed2 = parseFloat(token.actor.data.data.attributes.speed.otherSpeeds[i].total)}
 	}
 };
 return {land: land2, fly: fly2, swim: swim2, burrow: burrow2, climb: climb2, misc: miscSpeed2}
